feat(demo): keep focus in device view
